+++ src/lib.rs
@@ -5,6 +5,7 @@
 use bevy::prelude::*;
 
 pub(crate) mod animation;
+pub(crate) mod builder;
 pub(crate) mod error;
 pub(crate) mod layers;
 pub(crate) mod loader;
@@ -20,10 +21,11 @@ pub mod prelude {
         ManualTick, NextFrame, PlayDirection, RenderAnimation, emit_animation_frame_changed,
         render_children_animation, resolve_frame,
     };
+    pub use crate::builder::AsepriteBuilder;
     pub use crate::error::AsepriteError;
     pub use crate::layers::{
         AseFlip, AseRenderTarget, AseTexture, LayerEntry, LayerFilter, LayerId, SliceId,
-        SpriteLayerOf, SpriteLayers,
+        SpriteLayerOf, SpriteLayers, TagId,
     };
     pub use crate::loader::{
         Aseprite, AsepriteLoader, AsepriteLoaderPlugin, AsepriteLoaderSettings, SliceKeyMeta,